typescript refactor: gatsby-site
