+++ gatsby-site/gatsby-config.js
@@ -37,6 +37,15 @@ module.exports = {
     },
     `gatsby-plugin-styled-components`,
     'gatsby-plugin-netlify',
+    `gatsby-plugin-typescript`,
+    // {
+    //   resolve: `gatsby-plugin-typescript`,
+    //   options: {
+    //     isTSX: true, // defaults to false
+    //     jsxPragma: `jsx`, // defaults to "React"
+    //     allExtensions: true, // defaults to false
+    //   },
+    // },
     // this (optional) plugin enables Progressive Web App + Offline functionality
     // To learn more, visit: https://gatsby.dev/offline
     // `gatsby-plugin-offline`,
