+++ gatsby-site/jest.config.js
@@ -1,15 +1,17 @@
 module.exports = {
-  transform: {
-    '^.+\\.jsx?$': `<rootDir>/jest-preprocess.js`,
-  },
   moduleNameMapper: {
     '.+\\.(css|styl|less|sass|scss)$': `identity-obj-proxy`,
     '.+\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$': `<rootDir>/__mocks__/file-mock.js`,
   },
-  testPathIgnorePatterns: [`node_modules`, `.cache`, `public`],
   transformIgnorePatterns: [`node_modules/(?!(gatsby)/)`],
-  globals: {
-    __PATH_PREFIX__: ``,
-  },
   setupFiles: [`<rootDir>/loadershim.js`],
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  preset: 'ts-jest',
+  testEnvironment: 'jsdom',
+  testMatch: ['**/__tests__/**/*.ts?(x)', '**/?(*.)+(test).ts?(x)'],
+  testPathIgnorePatterns: [`node_modules`, `.cache`, `public`],
+  transform: {
+    '^.+\\.jsx?$': `<rootDir>/jest-preprocess.js`,
+    '^.+\\.ts?$': 'babel-jest',
+  },
 };
