+++ gatsby-site/src/components/DonationForm.tsx
@@ -1,37 +1,35 @@
-import React from 'react';
+import * as React from 'react';
 import { Formik, Field } from 'formik';
 import * as yup from 'yup';
+
+import DonateButton from './DonateButton';
 import FieldText from './FieldText';
 import Label from './Label';
-import DonateButton from './DonateButton';
 import DownArrow from './Form/DownArrow';
 import { FormError } from './Form/FormError';
 
-const SponsorFormSchema = yup.object({
-  company: yup
+const DonationFormSchema = yup.object({
+  firstName: yup
     .string()
     .trim()
-    .required('Please enter your company name'),
-  firstName: yup.string().trim(),
+    .required('Please enter your first name'),
   lastName: yup.string().trim(),
   email: yup
     .string()
     .trim()
-    .email()
+    .email('Please enter a valid email address')
     .required('Please enter your email'),
-  pledgeAmount: yup
+  monthlyPledge: yup
     .number()
-    .integer('Pledge must be an integer value')
-    .moreThan(0, 'Please enter a valid pledge amount')
-    .required('Please enter a pledge amount'),
+    .integer()
+    .moreThan(0, 'Please select a pledge tier'),
   pledgeDuration: yup.number().moreThan(0, 'Please select a pledge duration'),
 });
 
-const SponsorshipForm = ({ onSubmit }) => {
+const DonationForm = ({ onSubmit, ethPrices }) => {
   function handleDonate(values, actions) {
-    // console.log('submit', values);
+    // console.log('DonationForm:', 'submit', values);
 
-    // pass values to the callback
     onSubmit(values, actions);
     actions.setSubmitting(false);
   }
@@ -39,35 +37,32 @@ const SponsorshipForm = ({ onSubmit }) => {
   return (
     <Formik
       initialValues={{
-        company: '',
         firstName: '',
         lastName: '',
         email: '',
-        pledgeAmount: '',
+        monthlyPledge: 0,
         pledgeDuration: 0,
       }}
-      validationSchema={SponsorFormSchema}
+      validationSchema={DonationFormSchema}
       onSubmit={handleDonate}
     >
       {({ values, handleSubmit, handleChange, isSubmitting }) => (
-        <form onSubmit={handleSubmit} className="w-80-l w-90-m w-100 center" name="sponsorship-pledge">
-          <FieldText
-            type="text"
-            name="company"
-            id="pledge-company"
-            label="Company"
-            placeholder="Enter the name of your company"
-            value={values.company}
-            required
-          />
-
+        <form
+          data-testid="donation-form"
+          onSubmit={handleSubmit}
+          className="w-80-l w-90-m w-100 center"
+          name="donation-pledge"
+        >
           <FieldText
             type="text"
             name="firstName"
             id="pledge-first-name"
             label="First Name"
+            required
             placeholder="Enter your first name"
             value={values.firstName}
+            onChange={handleChange}
+            className="f6 input-reset b--black-10 pv3 ph2 db center w-100 br3 mt2"
           />
 
           <FieldText
@@ -78,6 +73,7 @@ const SponsorshipForm = ({ onSubmit }) => {
             placeholder="Enter your last name"
             value={values.lastName}
             onChange={handleChange}
+            className="f6 input-reset b--black-10 pv3 ph2 db center w-100 br3 mt2"
           />
 
           <FieldText
@@ -85,24 +81,35 @@ const SponsorshipForm = ({ onSubmit }) => {
             name="email"
             id="pledge-email"
             label="Email"
+            required
             placeholder="Enter your email address"
             value={values.email}
             onChange={handleChange}
-            required
+            className="f6 input-reset b--black-10 pv3 ph2 db center w-100 br3 mt2"
           />
 
-          <FieldText
-            type="number"
-            name="pledgeAmount"
-            id="pledge-amount"
-            label="Monthly Pledge Amount (USD)"
-            placeholder="Enter your pledge amount"
+          <Label required>Pledge Tier</Label>
+          <FormError name="monthlyPledge" className="pt2" />
+          <Field
+            as="select"
+            name="monthlyPledge"
             required
-            value={values.pledgeAmount}
+            className="f6 input-reset b--black-10 pv3 ph2 db center w-100 br3 mt2 bg-white black-50"
+            value={values.monthlyPledge}
             onChange={handleChange}
-            min="0"
-            step="1"
-          />
+            id="pledge-tier-select"
+          >
+            <option disabled={false} defaultValue="0" value="0">
+              Select your pledge tier
+            </option>
+            <option value="5">{`Student — $5/month (${ethPrices.student} ETH)`}</option>
+            <option value="15">{`Gold — $15/month (${ethPrices.gold} ETH)`}</option>
+            <option value="50">{`Platinum — $50/month (${ethPrices.platinum} ETH)`}</option>
+            <option value="150">{`Diamond — $150/month (${ethPrices.diamond} ETH)`}</option>
+            <option value="500">{`Ether Advisor — $500/month (${ethPrices.ether} ETH)`}</option>
+            <option value="1500">{`Elite Advisor — $1500/month (${ethPrices.elite} ETH)`}</option>
+          </Field>
+          <DownArrow />
 
           <Label required>How many months of your pledge will you prepay today?</Label>
           <FormError name="pledgeDuration" className="pt2" />
@@ -115,7 +122,7 @@ const SponsorshipForm = ({ onSubmit }) => {
             value={values.pledgeDuration}
             onChange={handleChange}
           >
-            <option disabled="" defaultValue="0" value="0">
+            <option disabled={false} defaultValue="0" value="0">
               Select the amount of months you would like to prepay for
             </option>
             <option value="1">1 month</option>
@@ -123,14 +130,13 @@ const SponsorshipForm = ({ onSubmit }) => {
             <option value="6">6 months</option>
             <option value="12">12 months</option>
           </Field>
-
           <DownArrow />
 
-          <DonateButton text="Sponsor" disabled={isSubmitting} />
+          <DonateButton disabled={isSubmitting} handleClick={handleSubmit} />
         </form>
       )}
     </Formik>
   );
 };
 
-export default SponsorshipForm;
+export default DonationForm;
