+++ gatsby-site/src/components/EventCard.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import styled from 'styled-components';
 import moment from 'moment';
 import Box from './system/Box';
