+++ gatsby-site/src/components/Form/DownArrow.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import arrowSvg from '../../img/arrow.svg';
 
 const DownArrow = () => {
