+++ gatsby-site/src/components/Form/FormError.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { ErrorMessage } from 'formik';
 
 export const FormError = props => {
