+++ gatsby-site/src/components/Identicon.tsx
