+++ gatsby-site/src/components/Input.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import styled from 'styled-components';
 import {
   space,
