+++ gatsby-site/src/components/Label.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import Required from './Required';
 
 const Label = props => {
