+++ gatsby-site/src/components/Layout.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import { ThemeProvider } from 'styled-components';
 
 // Global styles
