+++ gatsby-site/src/components/Required.tsx
@@ -1,3 +1,3 @@
-import React from 'react';
+import * as React from 'react';
 
 export default () => <b className="red f7"> *</b>;
