+++ gatsby-site/src/components/Section.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from 'react';
 import styled from 'styled-components';
 import { background, space, shadow, border, position, color } from 'styled-system';
