+++ gatsby-site/src/index.d.ts
@@ -1,0 +1,3 @@
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.svg';
